Plot of Metric MDS object - Shepard diagram

ggplotSampleMDSShepard uses ggplot2 to provide plot of Metric MDS results. Shepard diagram provides a scatter plot of :

on the x axis, the high dimensional pairwise distances between each sample pairs

on the y axis, the corresponding pairwise distances in the obtained low dimensional projection

ggplotSampleMDSShepard(
  mdsObj,
  nDim,
  title = "Multi Dimensional Scaling - Shepard's diagram",
  pointSize = 0.5,
  displayPseudoRSq = TRUE
)

Arguments

mdsObj

a MDS object calculated by the SMACOF algorithm using the computeMetricMDS() function

nDim

(optional) number of dimensions to use when calculating Shepard's diagram and pseudoRSquare. If missing, it will be set equal to the number of projection dimensions as calculated in mdsObj

title

title to give to the plot

pointSize

plot size of points

displayPseudoRSq

if TRUE, display pseudo RSquare in subtitle, on top of nb of dimensions

Value

a ggplot object

See also

Examples

# prepare data, build MDS object
example("computeMetricMDS") 
#> 
#> cmMMDS> library(CytoPipeline)
#> 
#> cmMMDS> data(OMIP021Samples)
#> 
#> cmMMDS> # estimate scale transformations 
#> cmMMDS> # and transform the whole OMIP021Samples
#> cmMMDS> 
#> cmMMDS> transList <- estimateScaleTransforms(
#> cmMMDS+     ff = OMIP021Samples[[1]],
#> cmMMDS+     fluoMethod = "estimateLogicle",
#> cmMMDS+     scatterMethod = "linearQuantile",
#> cmMMDS+     scatterRefMarker = "BV785 - CD3")
#> 
#> cmMMDS> OMIP021Trans <- CytoPipeline::applyScaleTransforms(
#> cmMMDS+     OMIP021Samples, 
#> cmMMDS+     transList)
#> 
#> cmMMDS> ffList <- flowCore::flowSet_to_list(OMIP021Trans)
#> 
#> cmMMDS> # As there are only 2 samples in OMIP021Samples dataset,
#> cmMMDS> # we create artificial samples that are random combinations of both samples
#> cmMMDS> 
#> cmMMDS> for(i in 3:5){
#> cmMMDS+     ffList[[i]] <- 
#> cmMMDS+         CytoPipeline::aggregateAndSample(
#> cmMMDS+             OMIP021Trans,
#> cmMMDS+             seed = 10*i,
#> cmMMDS+             nTotalEvents = 5000)[,1:22]
#> cmMMDS+ }
#> 
#> cmMMDS> fsNames <- c("Donor1", "Donor2", paste0("Agg",1:3))
#> 
#> cmMMDS> names(ffList) <- fsNames
#> 
#> cmMMDS> fsAll <- as(ffList,"flowSet")
#> 
#> cmMMDS> flowCore::pData(fsAll)$type <- factor(c("real", "real", rep("synthetic", 3)))
#> 
#> cmMMDS> flowCore::pData(fsAll)$grpId <- factor(c("D1", "D2", rep("Agg", 3)))
#> 
#> cmMMDS> # calculate all pairwise distances
#> cmMMDS> 
#> cmMMDS> pwDist <- pairwiseEMDDist(fsAll, 
#> cmMMDS+                              channels = c("FSC-A", "SSC-A"),
#> cmMMDS+                              verbose = FALSE)
#> 
#> cmMMDS> # compute Metric MDS object with explicit number of dimensions
#> cmMMDS> mdsObj <- computeMetricMDS(pwDist, nDim = 4, seed = 0)
#> 
#> cmMMDS> # mdsObj$nDim should be 4
#> cmMMDS> 
#> cmMMDS> #' # compute Metric MDS object by reaching a target pseudo RSquare
#> cmMMDS> mdsObj2 <- computeMetricMDS(pwDist, seed = 0, targetPseudoRSq = 0.999)

# Shepard diagrams 

p2D <- ggplotSampleMDSShepard(
    mdsObj,
    nDim = 2,
    pointSize = 1,
    title = "Shepard with 2 dimensions")

p3D <- ggplotSampleMDSShepard(
    mdsObj,
    nDim = 3,
    title = "Shepard with 3 dimensions") 
    #' 
pDefD <- ggplotSampleMDSShepard(
    mdsObj,
    title = "Shepard with default nb of dimensions")