Plot of Metric MDS object

ggplotSamplesMDS uses ggplot2 to provide plots of Metric MDS results. If both projectionAxes are in c(1,2), a global RSquare for these 2 axes is provided. For additional dimensions, a marginal RSquare per dimension is provided.

ggplotSamplesMDS(
  mdsObj,
  pData = data.frame(sampleId = seq_len(nrow(mdsObj$proj))),
  projectionAxes = c(1, 2),
  biplot = FALSE,
  extVariables = NULL,
  pDataForColour = NULL,
  pDataForShape = NULL,
  pDataForLabel = "name",
  pDataForAdditionalLabelling = NULL,
  sizeReflectingStress = FALSE,
  title = "Multi Dimensional Scaling",
  repelPointsLabels = TRUE,
  displayArrowLabels = TRUE,
  arrowThreshold = 0.7,
  repelArrowLabels = FALSE,
  flipXAxis = FALSE,
  flipYAxis = FALSE,
  ...
)

Arguments

mdsObj

a MDS object calculated by the SMACOF algorithm using the computeMetricMDS() function

pData

a data.frame providing user input sample data. These can be design of experiment variables, phenotype data per sample,... and will be used to highlight sample categories in the plot.

projectionAxes

which two axes should be plotted (should be a numeric vector of length 2)

biplot

if TRUE, adds projection of external variables

extVariables

are used to generate a biplot these are the external variables to regress with obtained configuration according to the two projection axes

pDataForColour

if not NULL, which pData variable will be used as colour aesthetic. Should be a character.

pDataForShape

if not NULL, which pData variable will be used as shape aesthetic. Should be a character.

pDataForLabel

if not NULL, which pData variable will be used as point labels in the plot. Should be a character.

pDataForAdditionalLabelling

if not NULL, which pData variable(s) will be add to the ggplot mapping, as to make them available for plotly tooltipping. Should be an array of character of maximum length 3. Note this works only if biplot=FALSE, as biplots contain circle and arrows that are currently not supported under ggplotly.

sizeReflectingStress

if TRUE, size of points will appear proportional to stress by point, i.e. the bigger the sample point appears, the less accurate its representation is (in terms of distances w.r.t. other points)

title

title to give to the plot

repelPointsLabels

if TRUE, uses ggrepel::geom_text_repel() instead of ggplot2::geom_text() (try to split the labels such that they do not overlap) for the points

displayArrowLabels

if TRUE, displays arrows labels (only with biplot)

arrowThreshold

(only with biplot), arrows will be made barely visible if their length is (in absolute value) less than this threshold.

repelArrowLabels

if TRUE, uses ggrepel::geom_text_repel() instead of ggplot2::geom_text() for the arrows (only with biplot)

flipXAxis

if TRUE, take the opposite of x values (provided as it might ease low dimensional projection comparisons)

flipYAxis

if TRUE, take the opposite of y values (provided as it might ease low dimensional projection comparisons)

...

additional parameters passed to ggrepel::geom_text_repel() (if used)

Value

a ggplot object

See also

Examples

# prepare data, build MDS object
example("computeMetricMDS")
#> 
#> cmMMDS> library(CytoPipeline)
#> 
#> cmMMDS> data(OMIP021Samples)
#> 
#> cmMMDS> # estimate scale transformations 
#> cmMMDS> # and transform the whole OMIP021Samples
#> cmMMDS> 
#> cmMMDS> transList <- estimateScaleTransforms(
#> cmMMDS+     ff = OMIP021Samples[[1]],
#> cmMMDS+     fluoMethod = "estimateLogicle",
#> cmMMDS+     scatterMethod = "linearQuantile",
#> cmMMDS+     scatterRefMarker = "BV785 - CD3")
#> estimating logicle transformations for fluorochrome channels...
#> Estimating linear transformation for scatter channels : reference marker = BV785 - CD3...
#> applying specific linear transformation for FSC-A channel...
#> initial quantiles : q5 = 10464.0482 ; q95 = 192509.607
#> target quantiles : q5 = 0.3849 ; q95 = 2.9395
#> a = 1.40e-05 ; b = 2.38e-01
#> applying FSC-A linear transformation for FSC-H channel...
#> applying specific linear transformation for SSC-A channel...
#> initial quantiles : q5 = 16345.6462 ; q95 = 245485.3555
#> target quantiles : q5 = 0.3849 ; q95 = 2.9395
#> a = 1.11e-05 ; b = 2.03e-01
#> applying SSC-A linear transformation for SSC-H channel...
#> 
#> cmMMDS> OMIP021Trans <- CytoPipeline::applyScaleTransforms(
#> cmMMDS+     OMIP021Samples, 
#> cmMMDS+     transList)
#> 
#> cmMMDS> ffList <- flowCore::flowSet_to_list(OMIP021Trans)
#> 
#> cmMMDS> # As there are only 2 samples in OMIP021Samples dataset,
#> cmMMDS> # we create artificial samples that are random combinations of both samples
#> cmMMDS> 
#> cmMMDS> for(i in 3:5){
#> cmMMDS+     ffList[[i]] <- 
#> cmMMDS+         CytoPipeline::aggregateAndSample(
#> cmMMDS+             OMIP021Trans,
#> cmMMDS+             seed = 10*i,
#> cmMMDS+             nTotalEvents = 5000)[,1:22]
#> cmMMDS+ }
#> 
#> cmMMDS> fsNames <- c("Donor1", "Donor2", paste0("Agg",1:3))
#> 
#> cmMMDS> names(ffList) <- fsNames
#> 
#> cmMMDS> fsAll <- as(ffList,"flowSet")
#> 
#> cmMMDS> # calculate all pairwise distances
#> cmMMDS> 
#> cmMMDS> pwDist <- getPairwiseEMDDist(fsAll, 
#> cmMMDS+                              channels = c("FSC-A", "SSC-A"),
#> cmMMDS+                              verbose = FALSE)
#> 
#> cmMMDS> # compute Metric MDS object with explicit number of dimensions
#> cmMMDS> mdsObj <- computeMetricMDS(pwDist, nDim = 4, seed = 0)
#> 
#> cmMMDS> # mdsObj$nDim should be 4
#> cmMMDS> 
#> cmMMDS> #' # compute Metric MDS object by reaching a target pseudo RSquare
#> cmMMDS> mdsObj2 <- computeMetricMDS(pwDist, seed = 0, targetPseudoRSq = 0.999)

# plot mds projection on axes 1 and 2,
# use 'group' for colour, 'type' for shape, and no label 

p_12 <- ggplotSamplesMDS(
    mdsObj = mdsObj,
    pData = flowCore::pData(fsAll),
    projectionAxes = c(1,2),
    pDataForColour = "group",
    pDataForLabel = NULL,
    pDataForShape = "type")

# plot mds projection on axes 3 and 4,
# use 'group' for colour, and 'name' as point label

p_34 <- ggplotSamplesMDS(
    mdsObj = mdsObj,
    pData = flowCore::pData(fsAll),
    projectionAxes = c(3,4),
    pDataForColour = "group",
    pDataForLabel = "name")

# plot mds projection on axes 1 and 2,
# use 'group' for colour, 'type' for shape, and 'name' as point label
# have sample point size reflecting 'stress'
# i.e. quality of projection w.r.t. distances to other points

p12_Stress <- ggplotSamplesMDS(
    mdsObj = mdsObj,
    pData = flowCore::pData(fsAll),
    projectionAxes = c(1,2),
    pDataForColour = "group",
    pDataForLabel = "name",
    pDataForShape = "type",
    sizeReflectingStress = TRUE)

# try to associate axes with median of each channel
# => use bi-plot

extVars <- getChannelsSummaryStat(
    fsAll,
    channels = c("FSC-A", "SSC-A"),
    statsFUN = stats::median)

bp_12 <- ggplotSamplesMDS(
    mdsObj = mdsObj,
    pData = flowCore::pData(fsAll),
    projectionAxes = c(1,2),
    biplot = TRUE,
    extVariables = extVars,
    pDataForColour = "group",
    pDataForLabel = NULL,
    pDataForShape = "type",
    seed = 0)

bp_34 <- ggplotSamplesMDS(
    mdsObj = mdsObj,
    pData = flowCore::pData(fsAll),
    projectionAxes = c(3,4),
    biplot = TRUE,
    extVariables = extVars,
    pDataForColour = "group",
    pDataForLabel = "name",
    seed = 0)